src/5. Tools and Agents/3_multi_agent_systems/3_Google_A2A_protocol/standalone agents/A2A_standalone_agent_weather/pyproject.toml
[project]
name = "a2a-sample-agent-langgraph"
version = "1.0.0"
description = "Sample LangGraph weather agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "jwcrypto>=1.5.6",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.16",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "pyowm>=3.3.0",
    "python-dotenv>=1.1.0",
    "setuptools>=80.4.0",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "tomli>=2.0.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
]
